requirements.txt
torch~=1.8.1
torchaudio~=0.8.1
torchtext~=0.9.0
pytorch-lightning~=1.3.0
hydra-core~=1.0
py7zr
folia
pynlpl
jupyter
matplotlib
rouge-metric
nltk
omegaconf
num2words
language_tool_python
spacy
torchlibrosa
